+++ src/main/kotlin/icu/windea/pls/config/core/config/ParadoxFileInfo.kt
@@ -9,12 +9,8 @@ class ParadoxFileInfo(
 	val rootInfo: ParadoxRootInfo
 ) {
 	val rootFile get() = rootInfo.rootFile
-	val descriptorFile get() = rootInfo.descriptorFile
-	val descriptorInfo get() = rootInfo.descriptorInfo
 	val rootType get() = rootInfo.rootType
 	val rootPath get() = rootInfo.rootPath
-	
-	val isValid get() = rootInfo.isValid 
 	
 	override fun equals(other: Any?): Boolean {
 		return this === other || other is ParadoxFileInfo && path == other.path && rootInfo == other.rootInfo