+++ src/main/kotlin/icu/windea/pls/config/core/config/ParadoxLauncherSettingsInfo.kt
@@ -1,8 +1,7 @@
 package icu.windea.pls.config.core.config
 
+import com.fasterxml.jackson.module.kotlin.*
 import com.intellij.openapi.vfs.*
-import icu.windea.pls.*
 import icu.windea.pls.core.*
-import icu.windea.pls.script.psi.*