+++ src/main/kotlin/icu/windea/pls/config/core/config/ParadoxRootInfo.kt
@@ -2,89 +2,135 @@ package icu.windea.pls.config.core.config
 
 import com.fasterxml.jackson.module.kotlin.*
 import com.intellij.openapi.vfs.*
-import com.intellij.psi.*
 import icu.windea.pls.*
 import icu.windea.pls.core.*
 import icu.windea.pls.script.psi.*
 import java.nio.file.*
 
 /**
- * @property rootFile 游戏或模组根目录
- * @property descriptorFile 描述符文件（`descriptor.mod`或`launcher-settings.json`）
- * @property markerFile 游戏类型标记文件（如`.stellaris`）
+ * @property rootFile 游戏或模组的根目录。
+ * @property gameRootFile 游戏文件的根目录。
  */
-class ParadoxRootInfo(
-	val rootFile: VirtualFile,
-	val descriptorFile: VirtualFile,
-	val markerFile: VirtualFile?,
-	val rootType: ParadoxRootType
-) {
-	val rootPath: Path = rootFile.toNioPath()
-	val gameTypeFromMarkerFile: ParadoxGameType? = markerFile?.let { file -> ParadoxGameType.resolve(file) }
-	val gameType: ParadoxGameType get() = gameTypeFromMarkerFile ?: getSettings().defaultGameType
-	val descriptorInfo: ParadoxDescriptorInfo? get() = doGetDescriptorInfo()
-	
-	val isValid get() = rootFile.isValid
-	val isAvailable get() = descriptorFile.isValid && (markerFile?.isValid != false)
-	
-	private fun doGetDescriptorInfo(): ParadoxDescriptorInfo? {
-		val result = descriptorFile.getUserData(PlsKeys.descriptorInfoKey)
-		if(result != null) return result
-		val resolved = runCatching { resolveDescriptorInfo(descriptorFile) }.getOrNull()
-		descriptorFile.putUserData(PlsKeys.descriptorInfoKey, resolved)
-		return resolved
-	}
-	
-	private fun resolveDescriptorInfo(descriptorFile: VirtualFile): ParadoxDescriptorInfo? {
-		val fileName = descriptorFile.name
-		return when {
-			fileName == PlsConstants.descriptorFileName -> {
-				val psiFile = descriptorFile.toPsiFile<PsiFile>(getDefaultProject())
-				if(psiFile !is ParadoxScriptFile) return null
-				val rootBlock = psiFile.findChild<ParadoxScriptRootBlock>() ?: return null
-				var name: String? = null
-				var version: String? = null
-				var picture: String? = null
-				var tags: Set<String>? = null
-				var supportedVersion: String? = null
-				var remoteFileId: String? = null
-				var path: String? = null
-				rootBlock.processProperty(includeConditional = false) { property ->
-					when(property.name) {
-						"name" -> name = property.findValue<ParadoxScriptString>()?.stringValue
-						"version" -> version = property.findValue<ParadoxScriptString>()?.stringValue
-						"picture" -> picture = property.findValue<ParadoxScriptString>()?.stringValue
-						"tags" -> tags = property.findBlockValues<ParadoxScriptString>().mapTo(mutableSetOf()) { it.stringValue }
-						"supported_version" -> supportedVersion = property.findValue<ParadoxScriptString>()?.stringValue
-						"remote_file_id" -> remoteFileId = property.findValue<ParadoxScriptString>()?.stringValue
-						"path" -> path = property.findValue<ParadoxScriptString>()?.stringValue
-					}
-					true
-				}
-				val nameToUse = name ?: descriptorFile.parent?.name.orAnonymous() //如果没有name属性，则使用根目录名
-				ParadoxDescriptorInfo(nameToUse, version, picture, tags, supportedVersion, remoteFileId, path, isModDescriptor = true)
-			}
-			
-			fileName == PlsConstants.launcherSettingsFileName -> {
-				val json = jsonMapper.readValue<Map<String, Any?>>(descriptorFile.inputStream)
-				val name = gameType.description
-				val version = json.get("rawVersion")?.toString()
-				ParadoxDescriptorInfo(name, version, isModDescriptor = false)
-			}
-			
-			else -> null
+sealed interface ParadoxRootInfo {
+    val rootFile: VirtualFile
+    val gameRootFile: VirtualFile
+    val rootType: ParadoxRootType
+    val gameType: ParadoxGameType
+    val isAvailable: Boolean
+    
+    val rootPath: Path
+    
+    companion object {
+        val values = mutableSetOf<ParadoxRootInfo>()
+    }
+}
+
+class ParadoxGameRootInfo(
+    override val rootFile: VirtualFile,
+    val launcherSettingsFile: VirtualFile,
+    override val rootType: ParadoxRootType,
+    val launcherSettingsInfo: ParadoxLauncherSettingsInfo,
+) : ParadoxRootInfo {
+    override val rootPath: Path = rootFile.toNioPath()
+	override val gameType: ParadoxGameType = launcherSettingsInfo.gameId.let { ParadoxGameType.resolve(it) } ?: throw IllegalStateException()
+    
+    override val gameRootFile: VirtualFile 
+		get() {
+			val dlcPath = launcherSettingsInfo.dlcPath
+			val path = launcherSettingsFile.toNioPath().resolve(dlcPath)
+			return VfsUtil.findFile(path, true) ?: throw IllegalStateException()
 		}
-	}
-	
-	override fun equals(other: Any?): Boolean {
-		return this === other || other is ParadoxRootInfo && rootFile == other.rootFile
-	}
-	
-	override fun hashCode(): Int {
-		return rootFile.hashCode()
-	}
-	
-	companion object {
-		val values = mutableSetOf<ParadoxRootInfo>()
-	}
+    
+    override val isAvailable get() = launcherSettingsFile.isValid
+    
+    override fun equals(other: Any?): Boolean {
+        return this === other || other is ParadoxRootInfo && rootFile == other.rootFile
+    }
+    
+    override fun hashCode(): Int {
+        return rootFile.hashCode()
+    }
+}
+
+class ParadoxLauncherSettingsInfo(
+    val gameId: String,
+    val version: String,
+    val rawVersion: String,
+    val gameDataPath: String, // %USER_DOCUMENTS%/Paradox Interactive/${gameDisplayName}
+    val dlcPath: String,
+    val exePath: String,
+    val exeArgs: List<String>
+) {
+    companion object Resolver {
+        fun resolve(file: VirtualFile): ParadoxLauncherSettingsInfo? {
+            try {
+                return jsonMapper.readValue(file.inputStream)
+            } catch(e: Exception) {
+                return null
+            }
+        }
+    }
+}
+
+class ParadoxModRootInfo(
+    override val rootFile: VirtualFile,
+    val descriptorFile: VirtualFile,
+    val markerFile: VirtualFile?,
+    override val rootType: ParadoxRootType,
+    val descriptorInfo: ParadoxDescriptorInfo
+) : ParadoxRootInfo {
+    override val rootPath: Path = rootFile.toNioPath()
+	override val gameType: ParadoxGameType = markerFile?.let { ParadoxGameType.resolve(it) } ?: getSettings().defaultGameType
+    override val gameRootFile: VirtualFile get() = rootFile
+    
+    override val isAvailable get() = descriptorFile.isValid && (markerFile?.isValid != false)
+    
+    override fun equals(other: Any?): Boolean {
+        return this === other || other is ParadoxRootInfo && rootFile == other.rootFile
+    }
+    
+    override fun hashCode(): Int {
+        return rootFile.hashCode()
+    }
+}
+
+class ParadoxDescriptorInfo(
+    val name: String,
+    val version: String? = null,
+    val picture: String? = null,
+    val tags: Set<String>? = null,
+    val supportedVersion: String? = null,
+    val remoteFileId: String? = null,
+    val path: String? = null,
+    val isModDescriptor: Boolean = true
+) {
+    //see: descriptor.cwt
+    
+    companion object Resolver {
+        fun resolve(file: VirtualFile): ParadoxDescriptorInfo? {
+            val psiFile = file.toPsiFile<ParadoxScriptFile>(getDefaultProject()) ?: return null
+            val rootBlock = psiFile.findChild<ParadoxScriptRootBlock>() ?: return null
+            var name: String? = null
+            var version: String? = null
+            var picture: String? = null
+            var tags: Set<String>? = null
+            var supportedVersion: String? = null
+            var remoteFileId: String? = null
+            var path: String? = null
+            rootBlock.processProperty(includeConditional = false) { property ->
+                when(property.name) {
+                    "name" -> name = property.findValue<ParadoxScriptString>()?.stringValue
+                    "version" -> version = property.findValue<ParadoxScriptString>()?.stringValue
+                    "picture" -> picture = property.findValue<ParadoxScriptString>()?.stringValue
+                    "tags" -> tags = property.findBlockValues<ParadoxScriptString>().mapTo(mutableSetOf()) { it.stringValue }
+                    "supported_version" -> supportedVersion = property.findValue<ParadoxScriptString>()?.stringValue
+                    "remote_file_id" -> remoteFileId = property.findValue<ParadoxScriptString>()?.stringValue
+                    "path" -> path = property.findValue<ParadoxScriptString>()?.stringValue
+                }
+                true
+            }
+            val nameToUse = name ?: file.parent?.name.orAnonymous() //如果没有name属性，则使用根目录名
+            return ParadoxDescriptorInfo(nameToUse, version, picture, tags, supportedVersion, remoteFileId, path, isModDescriptor = true)
+        }
+    }
 }