+++ src/main/kotlin/icu/windea/pls/core/library/ParadoxLibraryType.kt
@@ -7,6 +7,7 @@ import com.intellij.openapi.roots.libraries.ui.*
 import com.intellij.openapi.vfs.*
 import icons.*
 import icu.windea.pls.config.core.config.*
+import icu.windea.pls.core.*
 import javax.swing.*
 
 class ParadoxLibraryType : LibraryType<ParadoxLibraryProperties>(ParadoxLibraryKind) {
@@ -25,10 +26,9 @@ class ParadoxLibraryType : LibraryType<ParadoxLibraryProperties>(ParadoxLibraryK
 		if(dialog.showAndGet()) {
 			val rootFile = dialog.rootFile
 			val rootInfo = dialog.rootInfo
-			val descriptorInfo = rootInfo?.descriptorInfo
-			if(rootFile != null && rootInfo != null && descriptorInfo != null) {
+			if(rootFile != null && rootInfo != null) {
 				//基于descriptor.mod或launcher-settings.json得到库的名字，由于此文件可能发生变更，不保存库的属性
-				val libraryName = getLibraryName(rootInfo, descriptorInfo)
+				val libraryName = getLibraryName(rootInfo)
 				val libraryProperties = ParadoxLibraryProperties
 				return ParadoxNewLibraryConfiguration(libraryName, this, rootFile, libraryProperties)
 			}
@@ -36,14 +36,22 @@ class ParadoxLibraryType : LibraryType<ParadoxLibraryProperties>(ParadoxLibraryK
 		return null
 	}
 	
-	private fun getLibraryName(rootInfo: ParadoxRootInfo, descriptorInfo: ParadoxDescriptorInfo): String {
+	private fun getLibraryName(rootInfo: ParadoxRootInfo): String {
 		//FIXME 这里的名字和版本基于描述符文件，可能会过时
 		return buildString {
 			val rootType = rootInfo.rootType
 			append("Paradox/").append(rootInfo.gameType.description).append(" ").append(rootType.description)
-			val version = descriptorInfo.version
-			if(rootType == ParadoxRootType.Mod) append(": ").append(descriptorInfo.name)
-			if(version != null) append("@").append(version)
+			when(rootInfo) {
+				is ParadoxGameRootInfo -> {
+					val info = rootInfo.launcherSettingsInfo
+					append("@").append(info.version.escapeXml())
+				}
+				is ParadoxModRootInfo -> {
+					val info = rootInfo.descriptorInfo
+					append(": ").append(info.name.escapeXml())
+					if(!info.version.isNullOrEmpty()) append("@").append(info.version.escapeXml())
+				}
+			}
 		}
 	}
 }